Add ensure-bazelisk role
This adds a role which installs bazelisk if it is not already
present.  It also adds a test job which exercises the role; the
job will be run on any changes to the new role.

This also removes the noop job since it is no longer needed as
a placeholder.

Change-Id: I69b28b58338b46ec3fea696f9d97be4271ed6ae4