Refactor gerrit-base and gerrit-build by branch
Make gerrit-base include only the gerrit project, and jgit on
stable-3.1 and master, since that is required for plugin builds
as well.  This is the base job for all gerrit-related builds
(plugins and gerrit itself).

Then add a gerrit-build-base job which builds gerrit itself.
This holds the playbook for building gerrit and any other
configuration necessary, but it does not add the required-projects
(ie, the core plugins) because they are branch specific.

The gerrit-build job itself will be defined in the gerrit repo
but inherit from gerrit-build-base.  That job will have branch
variants for all supported gerrit branches, which the appropriate
list of required-projects for each.

A future change will add test jobs to this repo which will run
gerrit-base on changes to this repo to test changes to the main
part of the job.

Change-Id: Ib676f79593390d69935c1b9f5624b0c7fa1c6f90