Rename gerrit-base to gerrit-setup
Make gerrit-setup include only the gerrit project.  This will
run the playbook which sets up the git repos, but we will add
a new branch-aware job that inherits from this to the gerrit
repo to include the appropriate repos as required-projects.

Then add a gerrit-build-base job which builds gerrit itself.
This holds the playbook for building gerrit and any other
configuration necessary, but it does not add the required-projects
(ie, the core plugins) because they are also branch specific.

It is temporarily parented on gerrit-setup, but will be reparented
to gerrit-base once that job exists.

The gerrit-build job itself will be defined in the gerrit repo
but inherit from gerrit-build-base.  That job will have branch
variants for all supported gerrit branches, which the appropriate
list of required-projects for each.

A future change will add test jobs to this repo which will run
gerrit-base on changes to this repo to test changes to the main
part of the job.

Change-Id: Ib676f79593390d69935c1b9f5624b0c7fa1c6f90