Update patch set 8
Patch Set 8:

We could if we added opendev with the Gerrit driver, but since we used the git driver, it doesn't know how to parse that url, so we'll have to wait for it to land.  I say we go ahead and land this (even with the doc bug) and then remove it when the zuul-jobs change lands.

Patch-set: 8