Add gerrit build test jobs
Add several test jobs which build each supported branch of the
Gerrit repo to verify that they all work (since the Gerrit repo
itself is not running Zuul jobs yet).

Reparent gerrit-build and gerrit-build-plugin on gerrit-base,
now that it exists (it is defined by the dependent changes
below).  That job adds jgit to the branches which require it.

Depends-On: https://gerrit-review.googlesource.com/257073
Depends-On: https://gerrit-review.googlesource.com/257075
Depends-On: https://gerrit-review.googlesource.com/257076
Depends-On: https://gerrit-review.googlesource.com/257077
Change-Id: I4fd75dd043d5f883081920ce878e8686ff77e0d6