Correct some submodule edge cases
So that this role may be used for repos which both are not core
modules (and therefore do not have an entry in .gitmodules) and
also do not have matching branches, check to see if the repo
being processed is in .gitmodules.  If it is not, we should
perform the mv operation.

Also, handle the case where a repo is in .gitmodules but is not
a remote tracking branch.  In that case, we should perform the
submodule init operation to checkout the submodule ref, and we
should also add that case to the dependency chain warning check.

Finally, correct the destination directory in the submodule init
case (a typo had us using the wrong variable).

Change-Id: I1828dff9cfdbfd79da375141ea89e3b956f23190