Update patch set 4
Patch Set 4:

> mmm ... there is a lot more than a simple command I believe :-(

Yes, that's true, but I think we've already taken care of that.  There's a lot of stuff in there about installing and configuring java, and getting bazelisk.  But that's all done and we don't have to worry about it here.  We currently successfully *build* gerrit on all four branches.  The part missing is how to run the tests.  CI scripts or not, we should have a command to build gerrit, and a command to test.  Something like "bazelisk build ..." followed by "bazelisk test ...".

I just needed to know what's supposed to come after "bazelisk test", and I seem to have looked at the wrong source.  Hopefully the latest patchset corrects that.

> Let me see if I can change the scripts for making them NOT associated with Jenkins in any way.
> At the end of the day, we would need a build command to run locally as well for simulating what the CI is doing, so that in case of a failure people can run the same build locally.
> 
> > I've updated PS4 with the standalone strategy which I see in https://gerrit.googlesource.com/gerrit-ci-scripts/+/refs/heads/master/jenkins/gerrit-bazel-test.sh but I missed earlier.  I'm guessing that's the right script to be looking at for this job.
> 
> I believe there are more flags needed, I'll double-check.

Thanks!  If there's a better script to look at for the "run gerrit unit tests" job, please let me know.  :)

Patch-set: 4