Update patch set 5
Patch Set 5:

(3 comments)

I don't understand why the tests in the last patchset were failing, so to try to clear the situation up, I've temporarily transformed this job into a simple shell script which is a concatenation of the first two shell snippets from the gerrit-bazel-build template in the jenkins-ci-scripts.  Hopefully that will help us spot what's wrong.

Patch-set: 5