Update patch set 15
Patch Set 15:

The latest patchset takes what we learned in the previous several revisions and makes minimal changes to the existing job.  There is a new pre-run playbook to add the additional setup items needed for a full gerrit build, and we use the 16vcpu nodes for speed.

It looks like we're hitting a resource constraint in nodepool, so we may need to merge https://gerrit-review.googlesource.com/c/zuul/ops/+/261893 first.

Patch-set: 15