Bazel: Remove explicit java_toolchain option
Java toolchain is specified in .bazelrc file in gerrit workspace and
thus shouldn't be specified explicitly.

On master branch Bazel was updated to release 5.0, and the explicit
toolchain option is ignored anyway. On stable-3.5 branch, the code was
migrated to Java 11 language level, and hence toolchain_java8 was
removed and the build is failing for plugins on this branch.

Bug: Issue 15509
Change-Id: I97dc38f4bf19f57f84e24d850d22909e73f0293b