Add missing /usr/bin/python
maven_jar would fail to fetch material when a `python` executable is not
found in path. It looks it up with:

 python = ctx.which("python")

Since Debian 11 and Ubuntu 20.04 LTS, there is no more a
`/usr/bin/python` to get rid of the ambiguity between python 2 and
python 3. That can be set by installing `python-is-python3`.

Change-Id: I8a58778a78d4d65e2078bc24f0587519e91c70c3