bazelisk-build: Make test_output level variable
While seeing only errors might be nice, it's also nice to see all the
tests that ran, including passing ones, or even see all the test output,
depending on job preference. Add a variable for bazelisk_test_output
that uses the same values as specified on [1] to control the output
level.

[1] https://bazel.build/reference/command-line-reference#flag--test_output

Change-Id: Id773bfc4b3b1a6ee62150f2e1dbd62f638f78f82