Add .bazelrc and force color
Bazel automatically detects whether colors should be output, most
probably based on whether stdin is /dev/null or is a tty. Jenkins
supports ANSI coloring output so we can turn it on.

I did the same change to the Wikimedia container image we are using to
build Gerrit and its plugins:
https://gerrit.wikimedia.org/r/c/integration/config/+/809542

Change-Id: I05faadc4ce047271a954a45b2c04fc90c959c701