Do not attempt to map zuul/config as a gerrit submodule
The prepare-gerrit-repos role attempts to install any required-project
or the project of any change in a depends-on into gerrit as a submodule.
There are some cases where a change in a project may be in the dependency change but the project should not be treated as a submodule.
Since this is the exception, we special case those.

Add zuul/config as such a special case.

Change-Id: I80da22426ab62d55e114f1ef64937cf55306e9a0